+++ HackPhoneSrv/HackPhoneSrv/WndNumpad.h
@@ -11,7 +11,7 @@ protected:
 	DECLARE_MESSAGE_MAP()
 public:
 	CRect m_rc[15];
-	uint8_t m_cmd[15];
+	uint16_t m_cmd[15];
 	CPoint m_pt[15];
 public:
 	afx_msg void OnPaint();
